Consistently apply sanitization for all endpoints
